==> 1_get_sen_tweets.ipynb <==
import tweepy
import pandas as pd

infile = 'us_sen.csv'

with open(infile, 'r') as input:
    input.readline()
    raw_list = input.readlines()


sen_list = []
for line in raw_list:
    mini_list = line.strip().split(',')
    sen_list.append(mini_list)

ACCESS_TOKEN = ''
ACCESS_SECRET = ''
CONSUMER_KEY = ''
CONSUMER_SECRET = ''

# Setup access to API
def connect_to_twitter_OAuth():
    auth = tweepy.OAuthHandler(CONSUMER_KEY, CONSUMER_SECRET)
    auth.set_access_token(ACCESS_TOKEN, ACCESS_SECRET)

    api = tweepy.API(auth, wait_on_rate_limit=True, wait_on_rate_limit_notify=True)
    return api

# Create API object
api = connect_to_twitter_OAuth()

tweet_list = []
count = 1

for state, name, handle, party in sen_list:
    print(f"Working on {state}: {name} ==> {count}/{len(sen_list)}")
    temp_tweets = tweepy.Cursor(api.user_timeline, id=handle, tweet_mode='extended', since_id='1213578392492290048').items()

    tw_counter = 0
    for status in temp_tweets:
        tweet_list.append({'state': state, 'name': name, 'handle': handle, 'tweet': status.full_text, 'date':status.created_at,
                           'source': status.source, 'tweet_id': status.id,
                           'favorites': status.favorite_count, 'retweets': status.retweet_count,
                           'party': party})
        tw_counter += 1

    print(f'{state}:{name} done! >>> {tw_counter} tweets found')
    count += 1

df = pd.DataFrame(tweet_list, columns=['state', 'name', 'handle', 'tweet', 'date', 'source',
                                       'tweet_id', 'favorites', 'retweets', 'party'])

df.to_csv('raw_sen.csv')
