==> 3_sen models.ipynb <==
import json
import pandas as pd
import numpy as np

with open('sen_words_5000.json', 'r') as fp:
    sen_dict = json.load(fp)

df = pd.read_csv('raw_sen.csv', index_col=[0]).reset_index(drop=True)

all_words = list(sen_dict.keys())

##word to ID
wordtoID = {}
for i, word in enumerate(all_words):
    wordtoID[word] = i

bag = np.zeros((len(df), len(all_words)+1))

for i, tweet in enumerate(df['tweet']):
    for word in tweet.split():
        if word.strip().lower() in all_words:
            j = wordtoID[word.strip().lower()]
            bag[i][j] += 1
    if df['party'].iloc[i] == 'd':
        bag[i][-1] = 1

bag.shape

from sklearn.model_selection import train_test_split
X_train, X_test, y_train, y_test = train_test_split(bag[:,:-1], bag[:,-1], random_state=9)

##Logistic Regression

from sklearn.linear_model import LogisticRegression

lr = LogisticRegression()
lr.fit(X_train, y_train)

print(f'Testing score: {lr.score(X_test, y_test)}')
print(f'Training score: {lr.score(X_train, y_train)}')

##gaussian Naive Bayes

from sklearn.naive_bayes import GaussianNB
gnb = GaussianNB()
gnb.fit(X_train, y_train)

print(f'Testing score: {gnb.score(X_test, y_test)}')
print(f'Training score: {gnb.score(X_train, y_train)}')

from sklearn.ensemble import GradientBoostingClassifier
gbc = GradientBoostingClassifier()
gbc.fit(X_train, y_train)

print(f'Testing score: {gbc.score(X_test, y_test)}')
print(f'Training score: {gbc.score(X_train, y_train)}')
