==> 4_nltk models.ipynb <==
import nltk
import pandas as pd
from nltk.tokenize import word_tokenize
from nltk import pos_tag
from nltk.stem import PorterStemmer
from nltk.stem import WordNetLemmatizer 
from collections import defaultdict
import itertools
import numpy as np

with open('stopwords.txt') as infile:
    stop_words = infile.readlines()
    stop_words.extend(['rt', '&amp;'])

#returns a dictionary sorted  descending by its values
def sort_dict(d):
    return {k: v for k, v in sorted(d.items(), key=lambda item: item[1], reverse=True)}

#creates a dictionary of dictionaries that iterates through texts, counting the amount of words in each tweet per 'handle' or user
#returns a dictionary of dicitonaries
def create_dict(df, handles):
    
    big_dict = {}
    #currently retweets are treated as tweets from the senator
    for handle in handles: 
        handle_dict = defaultdict(int)
        for tweet in df[df['handle'] == handle]['tokens']:
            for word in tweet:
                handle_dict[word] += 1
        big_dict[handle] = sort_dict(handle_dict)
    return big_dict

#returns only the top 'n' amount of entries in a sorted dictionary
def dict_slice(d, n):
    d_sort = sort_dict(d)
    return dict(itertools.islice(d_sort.items(), n))

#processes a dictionary of dictionaries into one dictionary with total counts of all words
#returns a dictionary
def get_all_words(d):
    t_dict = defaultdict(int)
    for handle in d.keys():
        for word in d[handle].keys():
            t_dict[word] += d[handle][word]
    return sort_dict(t_dict)

def bag(series, words):
    mat = np.zeros((len(df), len(words)+1))
    
    wordtoID = {}
    for i, word in enumerate(words):
        wordtoID[word.strip().lower()] = i
    
    for i, tweet in enumerate(series):
        for word in tweet:
            if word.strip().lower() in words:
                j = wordtoID[word.strip().lower()]
                mat[i][j] += 1
        if df['party'].iloc[i] == 'd':
            mat[i][-1] = 1
            
    return mat

#converts a string into tokens, removes specific types of words based on part of speech, then stems for each token
#returns a list of stems
ps = PorterStemmer()
def get_tokens(str):
    tweet = ''
    for word in str.split():
        if not word.startswith('http') and word.lower() not in stop_words:
            tweet += word + ' '
    tokens = word_tokenize(tweet)
    pos = pos_tag(tokens)
    include_tags = ['VBN', 'VBD', 'JJ', 'JJS', 'JJR', 'CD', 'NN', 'NNS', 'NNP', 'NNPS']
    filtered_tokens = [tok for tok, tag in pos if tag in include_tags]
    return [ps.stem(tok).lower() for tok in filtered_tokens if len(tok) > 1]

df = pd.read_csv('raw_sen.csv', index_col=[0]).reset_index(drop=True)

df['tokens'] = df['tweet'].apply(get_tokens)

sen_tokens_dict = create_dict(df, df['handle'].unique())

all_words_d = get_all_words(sen_tokens_dict)

all_words_d_5000 = dict_slice(all_words_d, 5000)

bag_of_words = bag(df['tokens'], all_words_d_5000.keys())

bag_of_words.shape

from sklearn.model_selection import train_test_split
X_train, X_test, y_train, y_test = train_test_split(bag_of_words[:,:-1], bag_of_words[:,-1], random_state=9)

##gaussian Naive Bayes

from sklearn.naive_bayes import GaussianNB
gnb = GaussianNB()
gnb.fit(X_train, y_train)

print(f'Testing score: {gnb.score(X_test, y_test)}')
print(f'Training score: {gnb.score(X_train, y_train)}')

from sklearn.linear_model import LogisticRegression
lr = LogisticRegression()
lr.fit(X_train, y_train)

print(f'Testing score: {lr.score(X_test, y_test)}')
print(f'Training score: {lr.score(X_train, y_train)}')

lr1 = LogisticRegression(penalty='l1')
lr1.fit(X_train, y_train)

print(f'Testing score: {lr1.score(X_test, y_test)}')
print(f'Training score: {lr1.score(X_train, y_train)}')

from sklearn.discriminant_analysis import LinearDiscriminantAnalysis
lda = LinearDiscriminantAnalysis()
lda.fit(X_test, y_test)

print(f'Testing score: {lda.score(X_test, y_test)}')
print(f'Training score: {lda.score(X_train, y_train)}')
